requirements.txt
cached-property==1.4.3
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
djangorestframework==3.8.2
docker==3.4.1
docker-compose==1.22.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
idna==2.6
jsonschema==2.6.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
PyYAML==3.13
requests==2.18.4
six==1.11.0
texttable==0.9.1
urllib3==1.22
websocket-client==0.48.0
django-bootstrap3==11.0.0
celery==4.2.1
django-celery-results==1.0.1
django_celery_beat
redis
django-debug-toolbar